bpf: Pack ipv6_ct_tuple to match Golang
[ upstream commit b3759ca11d8976386e67e9ff518be15713c42bd8 ]

The Golang structure definition for ipv6_ct_tuple (the ct map key) would
not pad out the structure to the nearest 8B offset, which means that
when using Golang unsafe.Sizeof() for the userspace definition of this
struct, it would give a different size compared to the C sizeof() used
when declaring the CT map.

An upcoming commit relies on the userspace and kernel definitions for
these structures to have the same size, so this commit packs the ipv6
tuple to make them the same. Note that the ipv4_ct_tuple was already
marked as packed.

This should not have upgrade considerations because for Cilium 1.2, the
definition of the ct map values already changes so the map must be
deleted/recreated.

Signed-off-by: Joe Stringer <joe@covalent.io>
Signed-off-by: Thomas Graf <thomas@cilium.io>